Pulled out client v1 route logic into separate file
